Break out libdeflate_options from CustomMemoryAllocator
